% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_sigma_matrix.R
\name{compute_sigma_matrix}
\alias{compute_sigma_matrix}
\title{Compute Sigma Matrix}
\usage{
compute_sigma_matrix(time_series, p)
}
\arguments{
\item{time_series}{A numeric vector representing the time series data.}

\item{p}{An integer indicating the lag order to be used in the calculation.}
}
\value{
A covariance matrix of the given time series.
}
\description{
This function calculates the covariance matrix for a given time series with a specified lag order (p).
It employs the function of 'estimate_phi' and the variance at each level to construct
the covariance matrix in a band limited manner.
}
\examples{
ts_data <- rnorm(100) # Example time series data
cov_matrix <- compute_sigma_matrix(ts_data, 1) # Computing the covariance matrix

}
