% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_lagged_matrix.R
\name{create_lagged_matrix}
\alias{create_lagged_matrix}
\title{Create Lagged Matrix}
\usage{
create_lagged_matrix(time_series, p)
}
\arguments{
\item{time_series}{A numeric vector representing the time series data.}

\item{p}{An integer specifying the number of lags to create.}
}
\value{
A matrix with p columns, each representing a lagged version of the input time series.
        The matrix will have the same number of rows as the input time series, but the
        first p rows will contain NA values to account for the lag.
}
\description{
This function generates a matrix where each column is a lagged version of the input time series.
The number of lags (p) is specified by the user. NA values are used to fill in where data
is not available due to lagging, but the output will automatically remove rows are not complete.
}
\examples{
ts <- 1:10
create_lagged_matrix(ts, 2)

}
