% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ar_model.R
\name{estimate_ar_model}
\alias{estimate_ar_model}
\title{Estimate Autoregressive Model}
\usage{
estimate_ar_model(time_series, p, criterion, ARMA = T)
}
\arguments{
\item{time_series}{A numeric vector representing the time series data.}

\item{p}{An integer specifying the maximum lag order of the AR model.}

\item{criterion}{A string specifying the criterion for model selection. The valid options are "aic", "bic", "gbic", and "gaic".}
}
\value{
The value of the specified criterion for the estimated AR model.
}
\description{
This function estimates an autoregressive (AR) model using four different model selection criteria: AIC, BIC, generalized BIC (GBIC), and generalized BIC with prior information. The method for GBIC with prior information is based on the reference: https://doi.org/10.1111/rssb.12023.
}
\examples{
ts_data <- rnorm(100) # Example time series data
ar_model_aic <- estimate_ar_model(ts_data, 1, "aic") # Estimate using AIC
ar_model_bic <- estimate_ar_model(ts_data, 1, "bic") # Estimate using BIC
ar_model_gbic <- estimate_ar_model(ts_data, 1, "gbic") # Estimate using GBIC
ar_model_gbicp <- estimate_ar_model(ts_data, 1, "gbicp") # Estimate using GBICP

}
\references{
Zhao, Z., & Shao, X. (2013). Model selection for autoregressive processes with fractional time trends. Journal of the Royal Statistical Society: Series B (Statistical Methodology), 75(2), 323-341. https://doi.org/10.1111/rssb.12023
}
