% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_phi.R
\name{estimate_phi}
\alias{estimate_phi}
\title{Estimate Phi Function}
\usage{
estimate_phi(time_series, p, b)
}
\arguments{
\item{time_series}{A numeric vector representing the time series data.}

\item{p}{An integer specifying the number of lags to be used.}

\item{b}{A numeric value specifying the bandwidth for the local polynomial regression.}
}
\value{
A list containing three elements:
  - `phi_hat`: the estimated phi parameter,
  - `sigma_hat_2`: the estimated variance of the residuals,
  - `sigma_t_2`: the estimated variances at each time point.
}
\description{
This function estimates the phi parameter of a given time series. It uses a
local polynomial regression (loess) to estimate the variance of the residuals.
}
\examples{
ts_data <- rnorm(100)
estimate_phi(ts_data, p = 5, b = 0.5)
}
