% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_ts.R
\name{generate_ts}
\alias{generate_ts}
\title{Time Series Simulation Function}
\usage{
generate_ts(ar_coef = NULL, ma_coef = NULL, n, sigma2 = 1)
}
\arguments{
\item{ar_coef}{Vector of autoregressive coefficients. Default is \code{NULL}.}

\item{ma_coef}{Vector of moving average coefficients. Default is \code{NULL}.}

\item{n}{Length of the time series to generate.}

\item{sigma2}{Variance of the noise in the time series. Default is 1.}
}
\value{
A numeric vector representing the generated time series.
}
\description{
Generates a simulated time series based on specified autoregressive (AR), moving average (MA) coefficients, and other parameters.
}
\examples{
ts1 <- generate_ts(n = 100)
ts2 <- generate_ts(ar_coef = c(0.5, -0.2), n = 100)
ts3 <- generate_ts(ma_coef = c(0.5, 0.3), n = 100)
ts4 <- generate_ts(ar_coef = c(0.5, -0.2), ma_coef = c(0.5, 0.3), n = 100)
}
