% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterative_GLS.R
\name{iterative_GLS}
\alias{iterative_GLS}
\title{Iterative Generalized Least Squares Estimation for Time Series}
\usage{
iterative_GLS(
  Yt,
  epsilon = 1e-06,
  b = 0.04,
  max_iter = 1000,
  criterion = "gbicp"
)
}
\arguments{
\item{Yt}{A numeric vector representing the time series data.}

\item{epsilon}{A numeric value specifying the convergence threshold for the L2 norm of spline coefficients.}

\item{b}{bandwidth for kernel smoothing when estimating time varying variance of the time series.}

\item{max_iter}{An integer specifying the maximum number of iterations for the GLS estimation process.}

\item{criterion}{A string specifying the criterion for AR model selection. Valid options include "aic", "bic","gaic", "gbic", and "gbicp".}
}
\value{
A list containing various components of the GLS estimation process, including the number of iterations (`iter`),
        updated spline coefficients (`beta`), AR model coefficients (`phi`), detrended series (`mu`), inverse covariance matrix (`Sigma_inv`),
        and other AR model parameters.
}
\description{
This function performs an iterative generalized least squares (GLS) estimation on a given time series.
It starts by detrending the time series using a B-spline and ordinary least squares (OLS).
An autoregressive (AR) model is then fitted to the detrended series using a specified model selection criterion.
The covariance matrix of the detrended series is used to update the spline coefficients using GLS.
The process iteratively re-detrends the original time series until the L2 norm of the spline coefficients
between consecutive iterations is below a specified threshold or the maximum number of iterations is reached.
}
\examples{
ts_data <- rnorm(100) # Example time series data
gls_result <- iterative_GLS(ts_data) # Perform iterative GLS estimation

}
